requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# ── Telegram ───────────────────────────────────────────────────────────────────
telethon==1.37.0
python-telegram-bot==21.7

# ── AI / LLM ───────────────────────────────────────────────────────────────────
anthropic==0.40.0

# ── Task queue ─────────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.2.0

# ── Security ───────────────────────────────────────────────────────────────────
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib==1.7.4

# ── Config ─────────────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── Utils ──────────────────────────────────────────────────────────────────────
httpx==0.27.2
aiofiles==24.1.0
tenacity==9.0.0
